// src/payments/chargify/domain/types.ts
export type Subscription = {
  id: number;
  state: State;
  activated_at: string;
  balance_in_cents: string;
  cancel_at_end_of_period: string;
  canceled_at: string; // String date - Wed, 04 Oct 2023 12:09:32 EDT -04:00
  cancellation_message: string;
  coupon_code: string;
  created_at: string; // String date - Wed, 04 Oct 2023 12:09:32 EDT -04:00
  current_period_ends_at: string; // String date - Wed, 04 Oct 2023 12:09:32 EDT -04:00
  current_period_started_at: string; // String date - Wed, 04 Oct 2023 12:09:32 EDT -04:00
  delayed_cancel_at: string;
  expires_at: string;
  next_assessment_at: string;
  previous_state: string;
  payment_type: string;
  signup_payment_id: string;
  signup_revenue: string;
  total_revenue_in_cents: string;
  trial_ended_at: string;
  trial_started_at: string;
  updated_at: string;
  currency: string;
  customer: Customer;
  product: Product;
  credit_card: CreditCard;
  product_price_in_cents: number;
  last_payment_error?: LastPaymentError;
  name?: string;
  // to add dynamic properties
  [key: string]: any;
};
// ...
export type SubscriptionResponseObject = {
  subscription: Subscription;
};

export type CreateSubscriptionErrorResponseObject = {
  errors: Array<string>;
};

export type SubscriptionComponentResponseObject = {
  component: SubscriptionComponent;
};

export type LastPaymentError = {
  message?: string;
};

export type CreditCard = {
  customer_id: string;
  id: string;
  first_name: string;
  last_name: string;
  billing_address: string;
  billing_address_2: string;
  billing_city: string;
  billing_country: string;
  billing_state: string;
  billing_zip: string;
  card_type: string;
  current_vault: string;
  expiration_month: string;
  expiration_year: string;
  masked_card_number: string;
  vault_token: string;
  customer_vault_token: string;
};

export type Customer = {
  id: string;
  first_name: string;
  last_name: string;
  email: string;
  phone: string;
  address: string;
  address_2: string;
  city: string;
  country: string;
  created_at: Date;
  organization: string;
  reference: string;
  state: string;
  updated_at: Date;
  zip: string;
};

export type ProductFamily = {
  accounting_code: string;
  description: string;
  handle: string;
  id: string;
  name: string;
};

export type ProductUpdated = {
  id: number;
  product_price_point_id?: number;
  product_price_point_handle?: string;
  interval: number;
  interval_unit: string;
  name: string;
};

export type Product = {
  id: string;
  accounting_code: string;
  archived_at: Date;
  created_at: Date;
  description: string;
  expiration_interval: string;
  expiration_interval_unit: string;
  handle: string;
  initial_charge_in_cents: string;
  interval: string;
  interval_unit: string;
  name: string;
  price_in_cents: string;
  request_credit_card: string;
  require_credit_card: string;
  return_params: string;
  return_url: string;
  trial_interval: string;
  trial_interval_unit: string;
  trial_price_in_cents: string;
  update_return_url: string;
  updated_at: Date;
  use_site_exchange_rate: string;
  product_family: ProductFamily;
};

export type ProductResult = {
  product: Product;
};
// ...
interface Price {
  id: number;
  component_id: number;
  starting_quantity: number;
  ending_quantity: number | null;
  unit_price: string;
  price_point_id: number;
  formatted_unit_price: string;
  segment_id: number | null;
}

export class Component {
  id: number;
  name: string;
  handle: string;
  pricing_scheme: string;
  unit_name: string;
  unit_price: string;
  product_family_id: number;
  product_family_name: string;
  price_per_unit_in_cents: number | null;
  kind: string;
  archived: boolean;
  taxable: boolean;
  description: string;
  default_price_point_id: number;
  prices: Price[];
  price_point_count: number;
  price_points_url: string;
  default_price_point_name: string;
  tax_code: string;
  recurring: boolean;
  upgrade_charge: number | null;
  downgrade_credit: number | null;
  created_at: string;
  updated_at: string;
  archived_at: string | null;
  hide_date_range_on_invoice: boolean;
  allow_fractional_quantities: boolean;
  use_site_exchange_rate: boolean;
  item_category: string | null;
  accounting_code: string | null;
}

export class ComponentDto {
  component: Component;
}
// ...
export enum State {
  ACTIVE = 'active',
  CANCELED = 'canceled',
  EXPIRED = 'expired',
  ON_HOLD = 'on_hold',
  PAST_DUE = 'past_due',
  SOFT_FAILURE = 'soft_failure',
  TRIALING = 'trialing',
  TRIAL_ENDED = 'trial_ended',
  UNPAID = 'unpaid',
  SUSPENDED = 'suspended',
  AWAITING_SIGNUP = 'awaiting_signup',
}